Query count

+++ src/main/java/fiuba/tpp/reactorapp/model/response/ProcessCountResponse.java
@@ -5,17 +5,17 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class ProcessCountResponse {
 
     @JsonProperty("cantidadProcesos")
-    private Integer processCount;
+    private Long processCount;
 
-    public ProcessCountResponse(Integer processCount) {
+    public ProcessCountResponse(Long processCount) {
         this.processCount = processCount;
     }
 
-    public Integer getProcessCount() {
+    public Long getProcessCount() {
         return processCount;
     }
 
-    public void setProcessCount(Integer processCount) {
+    public void setProcessCount(Long processCount) {
         this.processCount = processCount;
     }
 }
